Removing useless constructor and unused imports
