+++ client/src/Components/Employees/EmployeePropertyCard.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { Segment, Button } from "semantic-ui-react";
-import { FlexRow, Container, FlexColumn } from "custom-components";
+import { Segment } from "semantic-ui-react";
 
 const EmployeePropertyCard = props => {
   return (
