+++ client/src/Components/Reservations/ReservationEdit.js
@@ -4,10 +4,7 @@ import {
   Header,
   Input,
   Button,
-  Label,
-  Statistic,
-  Popup,
-  Icon
+  Statistic
 } from "semantic-ui-react";
 import { FlexRow, FlexColumn, Text } from "custom-components";
 import { Link } from "react-router-dom";
